fix: context menu escape key press, multiple tooltip displayed together, stop propogating reaction action menu events
